backend/requirements.txt
fastapi
uvicorn
pillow
torch
torchvision
transformers
faiss-cpu
ultralytics
python-multipart
scikit-learn
numpy
opencv-python
mtcnn